# Optimize and fix