--- a/scr/AstarPathfinder/TempList.cs
+++ b/scr/AstarPathfinder/TempList.cs
@@ -1,4 +1,6 @@
-﻿namespace AstarPathfinder;
+﻿using System.Runtime.CompilerServices;
+
+namespace AstarPathfinder;
 
 using System;
 using System.Buffers;
@@ -23,15 +25,12 @@ internal ref struct TempList<T>
     {
         if(array.Length <= index)
         {
-            var newArray = ArrayPool<T>.Shared.Rent(index * 2);
-            Array.Copy(array, newArray, index);
-            ArrayPool<T>.Shared.Return(array, true);
-            array = newArray;
+            Resize();
         }
 
         array[index++] = item;
     }
-
+   
     public void RemoveAt(int i)
     {
         if ((uint)i >= (uint)this.index)
@@ -43,6 +42,23 @@ internal ref struct TempList<T>
             Array.Copy(array, i + 1, array, i, index - i);
         }
         index--;
+    }
+    
+    
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public void RemoveAtSwapBack(int i)
+    {
+        // ReSharper disable once LocalVariableHidesMember
+        var index = this.index;
+        if ((uint)i >= (uint)index)
+        {
+            ThrowArgumentOutOfRangeException(index, i);
+        }
+        if (i < index - 1)
+        {
+            array[i] = array[index - 1];
+        }
+        this.index--;
     }
 
     public ref T this[int i]
@@ -62,6 +78,16 @@ internal ref struct TempList<T>
     }
 
     public Span<T>.Enumerator GetEnumerator() => Span.GetEnumerator();
-
+    
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    private void Resize()
+    {
+        var newArray = ArrayPool<T>.Shared.Rent(index * 2);
+        Array.Copy(array, newArray, index);
+        ArrayPool<T>.Shared.Return(array, true);
+        array = newArray;
+    }
+    
+    [MethodImpl(MethodImplOptions.NoInlining)]
     private static void ThrowArgumentOutOfRangeException(int size, int i) => throw new ArgumentOutOfRangeException(nameof(i), $"Index {i} is out of range. Valid range: 0-{size - 1}");
 }