| Method        | Mean       | Error     | StdDev  | Allocated |
|-------------- |-----------:|----------:|--------:|----------:|
| Find          | 1,731.5 us | 145.81 us | 7.99 us |       4 B |
| FindOptimized |   315.4 us |  15.48 us | 0.85 us |         - |